Remove api server scaffolding
There will be no need for an api server if our content is (fairly)
static. We can host everything on a github page as long as the total
bandwidth per month is within 100gb. If our project is 2mb, that would
grant us a budget of 50,000 page views. This may be further optimized
with caching.