Add Tailwind CSS and shadcn/ui (#11)
Adds Tailwind CSS and shadcn to provide some easy styling and components
for this project! Most of the code was generated.

Note: I had to install shadcn with the `--legacy-peer-deps` flag, since
React 19 isn't included as a peer dependency. This may not be a problem
for us since for MVP it looks like the site will be pretty light, but a
good thing to note.